requirements.txt
selene>=2.0.0b5
pytest
allure-pytest